requirements.txt
appdirs==1.4.3
gevent==1.2.1
greenlet==0.4.12
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
